montana: Add qtiradio@1.0 to manifest
The manifest specifies version 2.1 but lshal outputs only @1.0. This
causes the following error in QTI's telephony framework:
07-15 07:35:55.218  1922  2092 E QTIRILJ : initQtiRadio: Exception: java.util.NoSuchElementException
[qti-telephony-common.jar:
com.qualcomm.qti.internal.telephony.QtiRIL.initQtiRadio -> HwBinder.getService("vendor.qti.hardware.radio.qtiradio@1.0::IQtiRadio")]
In addition, on deen's stock Q ROM [QPKS30.54-22-10] there are both @1.0
and @2.3 versions present, possibly as a result of commonizing of OEM's
device trees.
This might mean that the qcrild binary and the new libril-qc-hal-qmi.so
are meant to be used with qtiradio@2.X, and our legacy ril with
libril-qc-qmi-1.so is meant to use @1.0.
The vendor image is built to feature both sets for different devices,
with the modem being detected and the correct binaries ran at runtime
(init.class_main.sh).